Minor: Renaming files to the default application format
