+++ src/app/council-group/council-group.module.ts
@@ -3,7 +3,7 @@ import { Router, RouterModule } from '@angular/router';
 import { FormsModule } from '@angular/forms';
 import { CommonModule } from '@angular/common';
 
-import { CouncilGroupCreateComponent } from '../council-group/create/create-council-group.component';
+import { CouncilGroupCreateComponent } from '../council-group/create/council-group-create.component';
 import { CouncilGroupCreateService } from '../services/index';
 
 @NgModule({
